requirements.txt
astroid==2.5.1
beautifulsoup4==4.9.3
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
numpy==1.20.1
pandas==1.2.3
pdf2image==1.14.0
pdfminer.six==20201018
Pillow==8.1.2
pycparser==2.20
pylint==2.7.2
PyMuPDF==1.18.9
PyPDF2==1.26.0
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sortedcontainers==2.3.0
soupsieve==2.2
toml==0.10.2
wrapt==1.12.1
